♻️ refactor : 온보딩 헤더 파일 위치 변경
