+++ src/pages/mainPage/index.ts
@@ -1,4 +1,3 @@
 export * from "../onboarding/FirstMeet"
 export * from "../onboarding/MilitaryDay"
 export * from "../onboarding/CoupleContact"
-export * from "../../components/ui/progressheader"
