+++ src/pages/onboarding/Birthday.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router';
-import ProgressHeader from '@/components/ui/progressheader';
+import { ProgressHeader } from './components/progressheader';
 import { Button } from '@/components/ui/button';
 import Input from '@/components/ui/input';
 import { Checkbox } from '@/components/ui/checkbox';
